Added basic interp (Str, Num, Id, IfC)
Created an exprC type, don't know how/where/if we need to use it. Erlang is technically untyped (sans primatives).

Variables should have capital letters, types and records should be camel case;
atoms should all be lowercase.

I have Str in my Asgn3, not sure if it should be there, I never put it in interp so I think not...